Reconocimiento de manos

+++ A darle atomos/Assets/Scripts/IntArrayWrapper.cs
@@ -3,5 +3,5 @@ using System;
 [Serializable]
 public class IntArrayWrapper
 {
-    public int[] array;
+    public float[] array;
 }
